Changes back the color of the new links using "comment lines" to the "link color" for all themes (it was the same color as the titles).
Changes the debug theme to correctly use nesting (e.g. allowing to have bold or italic inside titles) and to use the same text styles it had before (underlined links, titles & numbers).

Changes all the screenshots to correctly reflect the changes both in the themes and the "tests.md" file.

Adds a link to the debug theme to the README.

Signed-off-by: Joris MASSON <masson.joris@gmail.com>